Added test script and a trained random forest model
In this commit,

I am using a test script that can be used to test the raw feature extraction with the help of a pre-trained random forest model on the raw features set presented in the repo.
Anyone can run the test script with giving the file path and program will load the trained model, extract the raw feature (53) and make a predication and output will be Benign file. or Malicious file. depending upon the result of perdication 0 or 1.